chore: rename packages, TS improvements, cleanup
