--- backend/.gitignore
@@ -1,7 +1,0 @@
-node_modules
-
-/.cache
-/build
-.env
-.dev.vars
-.wrangler
