+++ cdn/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "compilerOptions": {
+    "types": ["./worker-configuration.d.ts", "vite/client"]
+  }
+}
