+++ embed/package.json
@@ -1,16 +1,18 @@
 {
-  "name": "embed",
+  "name": "publisher-tools-embed",
   "version": "1.0.0",
   "description": "",
   "type": "module",
   "scripts": {
-    "build:dev": "vite build --config vite.config.ts --mode development",
-    "build": "vite build --config vite.config.ts"
+    "dev": "vite build --mode development",
+    "build": "vite build",
+    "prepare": "pnpm build"
   },
   "devDependencies": {
     "typescript": "^5.1.6",
     "vite": "^6.3.5"
   },
+  "files": ["dist/*"],
   "keywords": [],
   "author": "",
   "license": "ISC"
