--- frontend/.gitignore
@@ -1,8 +1,0 @@
-node_modules
-
-/.cache
-/build
-.env
-.dev.vars
-
-/public/init.js
