+++ frontend/app/routes/$type.import.tsx
@@ -1,5 +1,5 @@
-import { ImportModal } from '~/components/modals/index.js'
 import { useNavigate, useOutletContext } from '@remix-run/react'
+import { ImportModal } from '~/components/modals/index.js'
 import { APP_BASEPATH } from '~/lib/constants.js'
 import type { ElementConfigType } from '~/lib/types.js'
