+++ frontend/app/utils/sanitize.server.ts
@@ -1,6 +1,6 @@
 import sanitizeHtml from 'sanitize-html'
 import he from 'he'
-import type { SanitizedFields } from '../lib/types.js'
+import type { SanitizedFields } from '~/lib/types.js'
 
 export const sanitizeConfigFields = <T extends Partial<SanitizedFields>>(
   config: T
