+++ frontend/app/utils/validate.server.ts
@@ -4,7 +4,7 @@ import {
   PositionType,
   SlideAnimationType,
   WalletAddressFormatError
-} from '../lib/types.js'
+} from '~/lib/types.js'
 import { isWalletAddress, toWalletAddressUrl } from './utils.server.js'
 
 const rangeError = { message: 'Value has to be between 16 and 24' }
