+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@interledger/publisher-tools",
-  "version": "1.0.2",
+  "version": "0.0.0",
   "homepage": "https://github.com/interledger/publisher-tools#readme",
   "bugs": "https://github.com/interledger/publisher-tools/issues",
   "license": "Apache-2.0",
