+++ pnpm-workspace.yaml
@@ -1,5 +1,5 @@
 packages:
-  - 'backend'
+  - 'api'
   - 'frontend'
-  - 'worker'
   - 'embed'
+  - 'cdn'
