--- worker/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "types": ["./worker-configuration.d.ts"]
-  }
-}
